requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Deep learning
tensorflow>=2.8.0
torch>=1.11.0
xgboost>=1.5.0
lightgbm>=3.3.0

# Geospatial analysis
geopandas>=0.10.0
shapely>=1.8.0
folium>=0.12.0
geopy>=2.2.0
rasterio>=1.2.0

# Data processing
polars>=0.15.0
dask>=2022.1.0
pyarrow>=6.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.5.0
bokeh>=2.4.0
dash>=2.0.0

# API and web
fastapi>=0.75.0
uvicorn>=0.17.0
pydantic>=1.9.0
requests>=2.27.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
redis>=4.1.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.19.0
click>=8.0.0
tqdm>=4.62.0
joblib>=1.1.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.6.0

# Code quality
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.931

# Jupyter
jupyter>=1.0.0
jupyterlab>=3.2.0
ipywidgets>=7.6.0

# Monitoring and logging
mlflow>=1.24.0
wandb>=0.12.0
loguru>=0.6.0

# Deployment
gunicorn>=20.1.0
docker>=5.0.0
kubernetes>=21.7.0